Move database initialization and update into a separate package.
