+++ src/main/java/ro/sellfluence/db/versions/EmagMirrorDBVersion12.java
@@ -1,9 +1,9 @@
-package ro.sellfluence.db;
+package ro.sellfluence.db.versions;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 
-import static ro.sellfluence.db.EmagMirrorDBVersion1.executeStatement;
+import static ro.sellfluence.db.versions.EmagMirrorDBVersion1.executeStatement;
 import static ro.sellfluence.support.Time.timeE;
 
 class EmagMirrorDBVersion12 {
