+++ src/main/java/ro/sellfluence/db/versions/EmagMirrorDBVersion4.java
@@ -1,4 +1,4 @@
-package ro.sellfluence.db;
+package ro.sellfluence.db.versions;
 
 import java.sql.Connection;
 import java.sql.SQLException;
